No Google | using image-charts to get SVG and PNG
Added image-charts to get SVG and PNG Download.
Simply added this in Custom CSS too because of the new ID:

#qrcode2, #qrcode3 {
	display: flex;
	display: flex;
	justify-content: center;
	align-items: center;
	flex-direction: column;
}

But I can't fix my Share problem on Main-Adminpage with share button. The QR-Codes don't appear right because of Ajax!?. 
It seems to be loading in the wrong order, but I can't find the hook. The QR-Codes can correctly be downloaded at the Statistic Button Share Box. At the moment, I fixed it temporary with hiding the QR via CSS on Main-Adminpage:

.index.desktop #qrcode2, .index.desktop #qrcode3 {
	display: none !important;
}